fix(build): fix PrintVersionInfoTask to retrieve application label from merged manifest and resource files
